+++ src/xyz/oihana/schema/constants/traits/business/documents/GoodsReceiptConfirmationTrait.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\business\documents;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\business\documents\GoodsReceiptConfirmation} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\business\documents
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.3.0
+ */
+trait GoodsReceiptConfirmationTrait
+{
+    const string LINES                    = 'lines' ;
+    const string REFERENCES_DELIVERY_NOTE = 'referencesDeliveryNote' ;
+}