+++ src/xyz/oihana/schema/constants/traits/business/documents/GoodsReceiptLineTrait.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace xyz\oihana\schema\constants\traits\business\documents;
+
+/**
+ * The property name constants of the {@see \xyz\oihana\schema\business\documents\GoodsReceiptLine} class.
+ *
+ * @package xyz\oihana\schema\constants\traits\business\documents
+ * @author  Marc Alcaraz (eKameleon)
+ * @since   1.3.0
+ */
+trait GoodsReceiptLineTrait
+{
+    const string CONDITION         = 'condition' ;
+    const string DISCREPANCY_NOTE  = 'discrepancyNote' ;
+    const string EXPECTED_QUANTITY = 'expectedQuantity' ;
+    const string ITEM              = 'item' ;
+    const string POSITION          = 'position' ;
+    const string RECEIVED_QUANTITY = 'receivedQuantity' ;
+}